requirements.txt
--index-url https://pypi.org/simple

wheel
python-json-logger
tabulate
typing-extensions
adafruit-circuitpython-neopixel
adafruit-blinka
rpi_ws281x
RPi.GPIO
